Remove python2 from httplib2 module
